add gl float texture extension check.

+++ lib/src/filament/filament/backend/src/opengl/OpenGLProgram.cpp
@@ -100,7 +100,7 @@ OpenGLProgram::OpenGLProgram(OpenGLDriver* gl, const Program& programBuilder) no
         for (GLuint binding = 0, n = (GLuint) uniformBlockInfo.size(); binding < n; binding++) {
             auto const& name = uniformBlockInfo[binding];
             if (!name.empty()) {
-                if ((int) gl->getShaderModel() >= (int) backend::ShaderModel::GL_ES_30) {
+                if (gl->getShaderModel() >= backend::ShaderModel::GL_ES_30) {
                     GLint index = glGetUniformBlockIndex(program, name.c_str());
                     if (index >= 0) {
                         glUniformBlockBinding(program, GLuint(index), binding);
@@ -206,7 +206,7 @@ void OpenGLProgram::updateSamplers(OpenGLDriver* gl) noexcept {
 
             const GLTexture* const UTILS_RESTRICT t = gl->handle_cast<const GLTexture*>(th);
             if (UTILS_UNLIKELY(t->gl.fence)) {
-                if ((int) gl->getShaderModel() >= (int) backend::ShaderModel::GL_ES_30) {
+                if (gl->getShaderModel() >= backend::ShaderModel::GL_ES_30) {
                     glWaitSync(t->gl.fence, 0, GL_TIMEOUT_IGNORED);
                     glDeleteSync(t->gl.fence);
                 }
